implemented --nummon switch for number of monsters

+++ dungeon.h
@@ -60,7 +60,7 @@ typedef struct dungeon {
   uint8_t pc_distance[DUNGEON_Y][DUNGEON_X];
   uint8_t pc_tunnel[DUNGEON_Y][DUNGEON_X];
   pc_t pc;
-  int num_monsters; // assigning number of monsters
+  uint16_t num_monsters;
 } dungeon_t;
 
 void init_dungeon(dungeon_t *d);
